Init the framework of Node and Net protocol
This patch implement the basic framework of Node management
and network protocol, it also defined kinds of channels used
for internal modules' communication

There are bunch of change in there, we haven't split them as
small patch set cause this is a all-in-one initialization draft.
Meanwhile there are tons bugs in there as it hasn't be fully
tested. will improve in the later patch set.

Major change as below:
Support the p2p handshake protocol
Maintain the nodes list management
Utilization the message handle table to drive the node state
Define the channel for internal modules communication
Other tiny change like README, TODO

Signed-off-by: Yanbo Li <dreamfly281@gmail.com>