Complete EncodePoint function and add an error returned value.
The function of "EncodePoint" is used for compressing PublicKey
for less space used as same as which in bitcoin, and the process
is as follow:
If the choice of "isCompressed" is true, "EncodePoint" will reverse
X's value of PublicKey, and judge the value of Y. If Y is an even
number, a byte of "0x02" will be added before the value of X which
is already reversed, otherwise, 0x03 instead.
If the choice of "isCompressed" is false, the return value will be:
0x04 + X' + Y' ( X' and Y' were obtained by reversing the value of
X and Y).

Signed-off-by: XiaoJie Guo <guoxiaojie@onchain.com>