Fix that you must move the DNA project before it be compiled at local.
Fix that you must move the DNA project to "$GOPATH/src/github.com/
DNAProject/" before it be compiled at local from the patch of "Add
Travis-CI support". When Travis-CI builds DNA project, it will move
the project to "$GOPATH/src/DNAProject/", so we modified the path of
"import" statement from "DNA/...." to " github.com/DNAProject/DNA/..."
in all go source file in the patch of "Add Travis-CI support". But now,
we found a better way: we retained the original "import" statement, and
execute a command to modify it in the Travis-CI script. so we donn't
have to move the project to "$GOPATH/src/DNAProject/" at local.

1.insert a command to .travis.yml
2.restore the "import" statement from "github.com/DNAProject/DNA/..." to
'import "DNA/...."' in all go source files

Signed-off-by: XiaoJie Guo <guoxiaojie@onchain.com>